requirements.txt
aiohttp==3.6.2
alabaster==0.7.12
alembic==0.9.3.dev0
altgraph==0.17
apturl==0.5.2
asgiref==3.3.4
asn1crypto==0.24.0
astroid==2.2.5
async-timeout==3.0.1
attrs==19.3.0
autobahn==20.4.3
Babel==2.7.0
backcall==0.1.0
bleach==3.1.5
blinker==1.4
Brlapi==0.6.6
certifi==2020.12.5
cffi==1.14.0
chardet==4.0.0
click==6.7
cloudpickle==1.2.1
colorama==0.3.7
command-not-found==0.3
cryptography==2.9.2
cupshelpers==1.0
decorator==4.4.2
defer==1.0.6
defusedxml==0.6.0
distro-info===0.18ubuntu0.18.04.1
Django==3.2
django-filter==2.4.0
django-views==1.0.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.4.0
docopt==0.6.2
docutils==0.14
EasyProcess==0.2.7
entrypoints==0.3
Flask==0.12.2
Flask-BabelEx==0.9.3
Flask-Compress==1.4.0
Flask-Gravatar==0.4.2
Flask-Login==0.4.0
Flask-Mail==0.9.1
Flask-Migrate==2.1.1
Flask-Paranoid==0.2.0
Flask-Principal==0.4.0
Flask-Security-Too==3.4.2
Flask-SQLAlchemy==2.1
Flask-WTF==0.14.2
gunicorn==20.1.0
httplib2==0.9.2
idna==2.10
idna-ssl==1.1.0
imagesize==1.1.0
importlib-metadata==1.6.0
ipykernel==5.3.0
ipython==7.14.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
isort==4.3.20
itsdangerous==0.24
jedi==0.17.0
jeepney==0.4
Jinja2==2.11.2
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.3
jupyter-console==6.1.0
jupyter-core==4.6.3
jupyter-server-proxy==1.5.0
keyring==19.0.2
keyrings.alt==3.0
language-selector==0.1
launchpadlib==1.10.6
lazr.restfulclient==0.13.5
lazr.uri==1.0.3
lazy-object-proxy==1.4.1
ldap3==2.4.1
louis==3.5.0
macaroonbakery==1.1.3
Mako==1.0.7
MarkupSafe==1.1.1
mccabe==0.6.1
mistune==0.8.4
multidict==4.7.6
nbconvert==5.6.1
nbformat==5.0.6
netifaces==0.10.4
notebook==6.0.3
numpy==1.18.4
numpydoc==0.9.1
oauth==1.0.1
olefile==0.45.1
packaging==20.4
pandocfilters==1.4.2
paramiko==2.0.0
parso==0.7.0
passlib==1.7.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==5.1.0
pipreqs==0.4.10
prometheus-client==0.8.0
prompt-toolkit==3.0.5
protobuf==3.0.0
psutil==5.6.3
psycopg2==2.8.6
ptyprocess==0.6.0
pyasn1==0.4.2
pycairo==1.16.2
pycodestyle==2.5.0
pycparser==2.20
pycrypto==2.6.1
pycups==1.9.73
pyflakes==2.1.1
Pygments==2.6.1
pygobject==3.26.1
pyinotify==0.9.6
PyJWT==2.0.1
pylint==2.3.1
pymacaroons==0.13.0
PyNaCl==1.1.2
pyOpenSSL==17.5.0
pyparsing==2.4.7
PyQt5==5.12.2
PyQt5-sip==4.19.17
PyQtWebEngine==5.12.1
pyRFC3339==1.0
pyrsistent==0.16.0
pyscreenshot==0.5.1
python-apt==1.6.5+ubuntu0.5
python-dateutil==2.8.1
python-debian==0.1.32
pytz==2021.1
pyxdg==0.25
PyYAML==3.12
pyzmq==19.0.1
QtAwesome==0.5.7
qtconsole==4.7.4
QtPy==1.9.0
reportlab==3.4.0
requests==2.25.1
requests-unixsocket==0.1.5
roman==2.0.0
rope==0.14.0
SecretStorage==3.1.1
Send2Trash==1.5.0
simpervisor==0.3
simplejson==3.13.2
six==1.15.0
snowballstemmer==1.2.1
speaklater==1.4
Sphinx==2.1.1
sphinxcontrib-applehelp==1.0.1
sphinxcontrib-devhelp==1.0.1
sphinxcontrib-htmlhelp==1.0.2
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.2
sphinxcontrib-serializinghtml==1.1.3
spyder==3.3.4
spyder-kernels==0.4.4
SQLAlchemy==1.1.11
sqlparse==0.4.1
sshtunnel==0.1.4
system-service==0.3
systemd-python==234
terminado==0.8.3
testpath==0.4.4
tornado==6.0.4
traitlets==4.3.3
txaio==20.4.1
typed-ast==1.4.0
typing-extensions==3.7.4.3
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.26.4
usb-creator==0.3.3
virtualenv==15.1.0
vpython==7.6.1
wadllib==1.3.2
wcwidth==0.1.9
webencodings==0.5.1
Werkzeug==0.16.0
whitenoise==5.2.0
widgetsnbextension==3.5.1
wrapt==1.11.1
WTForms==2.1
wurlitzer==1.0.3
xkit==0.0.0
yarg==0.1.9
yarl==1.4.2
zipp==3.1.0
zope.interface==4.3.2
